remove old playground
