+++ packages/sit-onyx/src/components/examples/GridPlayground/GridOverlay.vue
